backend/requirements.txt
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.133.1
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
idna==3.11
joblib==1.5.3
nltk==3.9.3
passlib==1.7.4
pypdf==6.9.2
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.2.19
SQLAlchemy==2.0.47
starlette==0.52.1
textblob==0.19.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
google-generativeai==0.7.2
pytest==8.4.1
httpx==0.28.1
